움직이는 동심원 구현

+++ SimpleGame/Renderer.h
@@ -52,6 +52,6 @@ private:
 
 	GLuint m_FSSandboxShader = 0;
 	GLuint m_FSSandboxVBO = 0;
-	GLuint m_FSSandboxTime = 0;
+	float m_FSSandboxTime = 0;
 };
